# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Larder API",
    "version": "0.0.0"
  },
  "paths": {
...
    "/plans": {
      "post": {
        "tags": [
          "plans"
        ],
        "summary": "Create Plan",
        "operationId": "createPlan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanOut"
                }
              }
            }
          },
          "400": {
            "description": "bad start date, or no plan fits (reasons)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problems"
                }
              }
            }
          },
          "409": {
            "description": "setup incomplete (goals, allergies, preferences) or goals need review",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problems"
                }
              }
            }
          },
          "503": {
            "description": "the planner found no plan in time",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problems"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/plans/current": {
      "get": {
        "tags": [
          "plans"
        ],
        "summary": "Current Plan",
        "description": "The most recently created plan, or null if there is none.",
        "operationId": "getCurrentPlan",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/PlanOut"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Getcurrentplan"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/plans/{plan_id}": {
      "get": {
        "tags": [
          "plans"
        ],
        "summary": "Get Plan",
        "operationId": "getPlan",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanOut"
                }
              }
            }
          },
          "404": {
            "description": "no such plan (or not yours)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/recipes/{recipe_id}": {
      "get": {
        "tags": [
          "recipes"
        ],
        "summary": "Get Recipe",
        "description": "A recipe with allergens and suitable diets derived the same way the planner derives them\n(never TheMealDB's category, which isn't a diet claim).",
        "operationId": "getRecipe",
        "parameters": [
          {
            "name": "recipe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Recipe Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDetail"
                }
              }
            }
          },
          "404": {
            "description": "no such recipe"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Allergen": {
        "type": "string",
        "enum": [
          "celery",
          "gluten",
          "crustaceans",
          "eggs",
          "fish",
          "lupin",
          "milk",
          "molluscs",
          "mustard",
          "tree_nuts",
          "peanuts",
          "sesame",
          "soy",
          "sulphites"
        ],
        "title": "Allergen"
      },
...
      "DayOut": {
        "properties": {
          "day": {
            "type": "integer",
            "title": "Day"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "meals": {
            "items": {
              "$ref": "#/components/schemas/MealOut"
            },
            "type": "array",
            "title": "Meals"
          },
          "totals": {
            "$ref": "#/components/schemas/MacrosOut"
          },
          "logged": {
            "items": {
              "$ref": "#/components/schemas/MealLogOut"
            },
            "type": "array",
            "title": "Logged"
          }
        },
        "type": "object",
        "required": [
          "day",
          "date",
          "meals",
          "totals",
          "logged"
        ],
        "title": "DayOut"
      },
      "Diet": {
        "type": "string",
        "enum": [
          "vegetarian",
          "vegan",
          "pescatarian"
        ],
        "title": "Diet"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MacrosOut": {
        "properties": {
          "kcal": {
            "type": "integer",
            "title": "Kcal"
          },
          "protein_g": {
            "type": "integer",
            "title": "Protein G"
          },
          "fat_g": {
            "type": "integer",
            "title": "Fat G"
          },
          "carbs_g": {
            "type": "integer",
            "title": "Carbs G"
          }
        },
        "type": "object",
        "required": [
          "kcal",
          "protein_g",
          "fat_g",
          "carbs_g"
        ],
        "title": "MacrosOut",
        "description": "Energy and macros as whole numbers (kcal, grams)."
      },
...
      "PlanOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "start": {
            "type": "string",
            "format": "date",
            "title": "Start"
          },
          "days": {
            "items": {
              "$ref": "#/components/schemas/DayOut"
            },
            "type": "array",
            "title": "Days"
          },
          "planner": {
            "type": "string",
            "title": "Planner"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "solve_ms": {
            "type": "integer",
            "title": "Solve Ms"
          },
          "currency": {
            "type": "string",
            "enum": [
              "GBP",
              "EUR",
              "USD"
            ],
            "title": "Currency"
          },
          "targets": {
            "$ref": "#/components/schemas/TargetsOut"
          },
          "week_totals": {
            "$ref": "#/components/schemas/MacrosOut"
          },
          "cost_minor": {
            "type": "integer",
            "title": "Cost Minor",
            "description": "estimated shopping cost"
          },
          "budget_minor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Minor"
          },
          "pantry_used_g": {
            "type": "integer",
            "title": "Pantry Used G"
          },
          "waste_g": {
            "type": "integer",
            "title": "Waste G",
            "description": "pantry food that will expire unused within the plan"
          },
          "shopping_items": {
            "type": "integer",
            "title": "Shopping Items"
          },
          "excluded_allergens": {
            "items": {
              "$ref": "#/components/schemas/Allergen"
            },
            "type": "array",
            "title": "Excluded Allergens"
          },
          "diet": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Diet"
              },
              {
                "type": "null"
              }
            ]
          },
          "violations": {
            "items": {
              "$ref": "#/components/schemas/ViolationOut"
            },
            "type": "array",
            "title": "Violations"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "version",
          "parent_id",
          "start",
          "days",
          "planner",
          "status",
          "created_at",
          "solve_ms",
          "currency",
          "targets",
          "week_totals",
          "cost_minor",
          "budget_minor",
          "pantry_used_g",
          "waste_g",
          "shopping_items",
          "excluded_allergens",
          "diet",
          "violations",
          "notes"
        ],
        "title": "PlanOut"
      },
      "PlanRequest": {
        "properties": {
          "start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start",
            "description": "first day (the user's local date, from yesterday to a week ahead); defaults to today (UTC)"
          }
        },
        "type": "object",
        "title": "PlanRequest"
      },
...
      "Problems": {
        "properties": {
          "detail": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "Problems",
        "description": "422 body for domain-rule failures (e.g. unsafe calorie targets)."
      },
...
      "RecipeDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "cuisine": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cuisine"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "instructions": {
            "type": "string",
            "title": "Instructions"
          },
          "generated": {
            "type": "boolean",
            "title": "Generated"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "servings": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Servings"
          },
          "servings_estimated": {
            "type": "boolean",
            "title": "Servings Estimated"
          },
          "per_serving": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MacrosOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "meal_types": {
            "items": {
              "$ref": "#/components/schemas/Slot"
            },
            "type": "array",
            "title": "Meal Types"
          },
          "allergens": {
            "items": {
              "$ref": "#/components/schemas/Allergen"
            },
            "type": "array",
            "title": "Allergens"
          },
          "allergens_complete": {
            "type": "boolean",
            "title": "Allergens Complete"
          },
          "suitable_for": {
            "items": {
              "$ref": "#/components/schemas/Diet"
            },
            "type": "array",
            "title": "Suitable For"
          },
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/RecipeIngredientOut"
            },
            "type": "array",
            "title": "Ingredients"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "category",
          "cuisine",
          "image_url",
          "instructions",
          "generated",
          "source",
          "source_url",
          "servings",
          "servings_estimated",
          "per_serving",
          "meal_types",
          "allergens",
          "allergens_complete",
          "suitable_for",
          "ingredients"
        ],
        "title": "RecipeDetail"
      },
...
      "RecipeIngredientOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "measure": {
            "type": "string",
            "title": "Measure"
          },
          "food_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Food Id"
          },
          "food_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Food Name"
          },
          "grams": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grams"
          }
        },
        "type": "object",
        "required": [
          "name",
          "measure",
          "food_id",
          "food_name",
          "grams"
        ],
        "title": "RecipeIngredientOut"
      },
...
      "Slot": {
        "type": "string",
        "enum": [
          "breakfast",
          "lunch",
          "dinner",
          "snack"
        ],
        "title": "Slot",
        "description": "A meal of the day, in eating order."
      },
...
      "TargetsOut": {
        "properties": {
          "daily": {
            "$ref": "#/components/schemas/NutrientBands"
          },
          "weekly": {
            "$ref": "#/components/schemas/NutrientBands"
          },
          "calorie_floor": {
            "type": "integer",
            "title": "Calorie Floor"
          }
        },
        "type": "object",
        "required": [
          "daily",
          "weekly",
          "calorie_floor"
        ],
        "title": "TargetsOut",
        "description": "What the planner aimed for: per-day bands, whole-plan bands, and the calorie floor."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ViolationOut": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "hard": {
            "type": "boolean",
            "title": "Hard"
          },
          "day": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day"
          },
          "slot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Slot"
              },
              {
                "type": "null"
              }
            ]
          },
          "recipe_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipe Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "hard"
        ],
        "title": "ViolationOut",
        "description": "A rule the plan misses. `hard` ones (kcal band, budget, repeats) come only from the\nbaseline planner; soft ones (protein/fat/carbs bands) are trade-offs any planner may make.\nSafety rules (allergens, diet, calorie floor) are never broken in a saved plan."
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}